dry run opts

+++ aip2b2.rb
@@ -8,6 +8,11 @@ ARGV.options do |opts|
   opts.parse!
 end
 
+puts $dryrun
+puts 'Class is:'
+puts $dryrun.class
+exit
+
 # Set up methods
 def premisreport(actiontype,outcome)
    @premis_structure['events'] << [{'eventType':actiontype,'eventDetail':$command,'eventDateTime':Time.now,'eventOutcome':outcome}] 
@@ -38,13 +43,19 @@ end
    b2_target = '' + packagename
    logfile = @target_path + 'data' + 'logs' + "#{packagename}.log"
    @premis_structure = JSON.parse(File.read(logfile))
-   $command = 'b2 sync ' + $dryrun + '"' + @target_path.to_s + '" ' + '"' + b2_target + '"'
+   if $dryrun.nil?
+    $command = 'b2 sync ' + '"' + @target_path.to_s + '" ' + '"' + b2_target + '"'
+  else
+    $command = 'b2 sync ' + $dryrun + '"' + @target_path.to_s + '" ' + '"' + b2_target + '"'
+  end
 
    if system($command)
-     puts "SUCCESS!".green
-     premisreport('replication','pass')
-     puts @premis_structure
-     File.open("#{@target_path.to_s}/#{packagename}_#{Time.now}_premis.log",'w') {|file| file.write(@premis_structure.to_json)}
-     system($command) 
+    puts "SUCCESS!".green
+    premisreport('replication','pass')
+    puts @premis_structure
+    if ! if $dryrun.nil?
+      File.open("#{@target_path.to_s}/#{packagename}_#{Time.now}_premis.log",'w') {|file| file.write(@premis_structure.to_json)}
+      system($command)
+    end
    end
  end
